Add controllers, services, and repositories for Cliente, Compra, Comuna, Detalle, Producto, Region, and Sucursal; update application properties for Hibernate validation

+++ src/main/java/com/ecomarket/ecomarket/model/Compra.java
@@ -20,7 +20,7 @@ public class Compra {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID_COMPRA")
-    private Long idCompra;
+    private int idCompra;
 
     @Column(name = "FECHA_COMPRA", nullable = false)
     private Date fechaCompra;
